[3] refactor: merge …/graph_links/{links, storage}.rs
