+++ lib/segment/src/index/hnsw_index/graph_links/mod.rs
@@ -19,7 +19,7 @@
 //!   ┌───────────────────────────────────────────────────────────────────────┐
 //!   │ GraphLinks  (links.rs) — public handle, a self_cell of:               │
 //!   │                                                                       │
-//!   │   owner:  GraphLinksEnum (storage.rs)   dependent: GraphLinksView<'_> │
+//!   │   owner:  GraphLinksEnum (links.rs)     dependent: GraphLinksView<'_> │
 //!   │           ├─ Ram(Vec<u8>)                          (view.rs)          │
 //!   │           └─ Universal(                  zero-copy parse that BORROWS │
 //!   │                Box<dyn GraphLinksStorage>)  the owner's bytes; every  │
@@ -34,8 +34,7 @@
 //!
 //! | file           | responsibility                                                       |
 //! |----------------|----------------------------------------------------------------------|
-//! | [`links`]      | [`GraphLinks`] — public handle pairing owned bytes with a parsed view |
-//! | [`storage`]    | `GraphLinksEnum` + `GraphLinksStorage`: RAM vs. universal-IO backing  |
+//! | [`links`]      | [`GraphLinks`] + `GraphLinksEnum` — public handle pairing owned bytes (RAM vs. universal-IO) with a parsed view |
 //! | [`view`]       | `GraphLinksView` — zero-copy parser/accessor over the bytes           |
 //! | [`serializer`] | [`serialize_graph_links`] — edges → serialized bytes                  |
 //! | [`header`]     | on-disk header structs shared by the serializer and the view         |
@@ -78,7 +77,6 @@ mod format;
 mod header;
 mod links;
 mod serializer;
-mod storage;
 mod vectors;
 mod view;
